decompile: upstream b00731fc12cbd7aa49c00b5828ee468bbd51c9b3 (2023-12-09) -- 30 files changed
Upstream commit: https://github.com/espressif/esp-thread-lib/commit/b00731fc12cbd7aa49c00b5828ee468bbd51c9b3

Upstream subject: feat(br): update border router lib     esp-openthread: 8d18b44     openthread: 41ef807

Archives decompiled: 1

+++ decompiled/libopenthread_br/esp_openthread_discovery.cpp/discovery_delegate_update.c
@@ -1,8 +1,8 @@
 /*
- * Last changed at upstream commit 8d47b585cf4b5e717aa06b2897d27c843fafa343
- * https://github.com/espressif/esp-thread-lib/commit/8d47b585cf4b5e717aa06b2897d27c843fafa343
- * Upstream date: 2022-05-24 22:56:58 +0800
- * Upstream subject: openthread: rebuild the lib with new toolchain
+ * Last changed at upstream commit b00731fc12cbd7aa49c00b5828ee468bbd51c9b3
+ * https://github.com/espressif/esp-thread-lib/commit/b00731fc12cbd7aa49c00b5828ee468bbd51c9b3
+ * Upstream date: 2023-12-09 16:01:37 +0800
+ * Upstream subject: feat(br): update border router lib     esp-openthread: 8d18b44     openthread: 41ef807
  * Source: libopenthread_br -> esp_openthread_discovery.cpp.o -> discovery_delegate_update
  *
  * (C) Espressif, Apache License 2.0.
@@ -21,7 +21,7 @@ void discovery_delegate_update(esp_openthread_mainloop_context_t *param_1)
   
   uVar1 = s_mdns_event_fd;
   if (-1 < (int)s_mdns_event_fd) {
-    if ((int)s_mdns_event_fd < 0x40) {
+    if (s_mdns_event_fd < 0x40) {
       peVar3 = param_1 + ((int)s_mdns_event_fd >> 5) * 4;
       uVar2 = 1 << (s_mdns_event_fd & 0x1f);
       *(uint *)peVar3 = *(uint *)peVar3 | uVar2;
